This commit was generated by cvs2svn to compensate for changes in r5789, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@5790 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/c64tpi.c
@@ -31,6 +31,7 @@
 #include "maincpu.h"
 #include "tpi.h"
 #include "tpicore.h"
+#include "types.h"
 
 /*----------------------------------------------------------------------*/
 /* renaming of exported functions */
@@ -68,7 +69,7 @@ _TPI_FUNC void tpi_set_cb(int a)
 }
 
 static int ieee_is_dev = 1;
-static int ieee_is_out = 1;
+static BYTE ieee_is_out = 1;
 
 _TPI_FUNC void _tpi_reset(void)
 {
